rapport - algorithmes

+++ Services/Recuit.py
@@ -11,9 +11,8 @@ class Recuit:
         self.voisinage = voisinage
         self.fitness = fitness
     
-    def resolve(self, x: list, t: int, numberchangetemp: int, numbermovestemp: int, mu: float):
+    def resolve(self, x: list, t: float, numberchangetemp: int, numbermovestemp: int, mu: float):
         xmin = nextx = x
-        # todo update with real f
         fmin = fx = fnextx = self.fitness.calcul(x)
 
         # instanciate var
